remove appearance="strong", apply mobile-desktop sizings
