+++ packages/icon/index.ts
@@ -1,3 +1,4 @@
 export { Icon } from "./src/icon";
 
-export type { IconProps, SvgAttributes } from "./src/icon";
+export type { IconProps } from "./src/icon";
+export type { SvgAttributes } from "./src/types";
