+++ packages/icon/src/icon.tsx
@@ -1,7 +1,7 @@
-import { Children, cloneElement, HTMLAttributes, isValidElement, ReactElement, ReactNode } from "react";
-import invariant from "tiny-invariant";
+import { ReactNode } from "react";
 import { cx } from "../../cx";
 import { WithStyleProps } from "../../types";
+import { IconBase } from "./_icon-base";
 
 type IconProps = WithStyleProps & {
 	/**
@@ -9,11 +9,6 @@ type IconProps = WithStyleProps & {
 	 */
 	svg: ReactNode;
 };
-
-type SvgAttributes = HTMLAttributes<SVGElement> & {
-	focusable?: "true" | "false";
-};
-
 /**
  * Decorates an svg icon with automatic sizing styles.
  * Merges `className` selectors with the following order of precedence (last one wins):
@@ -21,19 +16,9 @@ type SvgAttributes = HTMLAttributes<SVGElement> & {
  * 2. svg className
  * 3. Icon className
  */
-const Icon = ({ className, style, svg }: IconProps) => {
-	const icon = Children.only(svg) as ReactElement;
-	invariant(isValidElement<SvgAttributes>(icon), "Icon must be passed a single SVG icon as a JSX tag.");
-
-	return (
-		<>
-			{cloneElement(icon, {
-				className: cx("size-6 sm:size-5 shrink-0", icon.props.className, className),
-				style: { ...icon.props.style, ...style },
-			})}
-		</>
-	);
-};
+const Icon = ({ className, style, svg }: IconProps) => (
+	<IconBase className={cx("size-6 sm:size-5", className)} style={style} svg={svg} />
+);
 
 export { Icon };
-export type { IconProps, SvgAttributes };
+export type { IconProps };
