+++ packages/icon/src/types.ts
@@ -1,0 +1,5 @@
+import type { HTMLAttributes } from "react";
+
+export type SvgAttributes = HTMLAttributes<SVGElement> & {
+	focusable?: "true" | "false";
+};
